Remove util.go and use functionality of exsits directly where needed
Having a util file can be considered bad practice as it does not
communicate anything about its content but the fact that those are
random function that do not seem to fit anywhere.
In this case especially the util file contained one method only which
was used in one place only. Also the only difference between the exist
method and _, err := os.Stat(...); !os.IsNotExist(err) is nothing but
one exclamation mark.